+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKCE2SS9M1NH5JJFYY38CVN1"
+        "01KKE4JZ72AV27CB4Q928X4GGK"
     )
     r"""The unique identifier of the data source"""